requirements.txt
alpha_vantage
beautifulsoup4
python-edgar
gensim
ipython
joblib
matplotlib
nltk
numpy
pandas
polyleven
python_dateutil
python_edgar
pytz
PyYAML
Quandl
requests
scikit_learn
sentence_transformers
torch
transformers
yfinance
top2vec[sentence_encoders]
tensorflow
tensorflow_hub
tensorflow_text
top2vec